feat: Model loading has been added

+++ include/IndexBuffer.h
@@ -6,12 +6,12 @@
 class IndexBuffer
 {
 public:
-	IndexBuffer(const ContentBufferCreateInfo<uint16_t>& contentBufferCreateInfo);
+	IndexBuffer(const ContentBufferCreateInfo<uint32_t>& contentBufferCreateInfo);
 	const size_t getIndexCount() const;
 	const vk::Buffer getVulkanBuffer() const;
 
 private:
-	void createIndexData(const std::vector<uint16_t>& indices, const PhysicalDeviceProperties& physicalDeviceProperties);
+	void createIndexData(const std::vector<uint32_t>& indices, const PhysicalDeviceProperties& physicalDeviceProperties);
 	size_t indexCount;
 	Buffer buffer;
 	StagingBuffer stagingBuffer;
